--- src/mock/index.ts
@@ -1,4 +1,0 @@
-export { CheerioServiceMock } from './CheerioServiceMock';
-export { generateMockPage } from './generateMockPage';
-export { generateMockUser } from './generateMockUser';
-export { PageRepositoryMock } from './PageRepositoryMock';
