+++ src/mock/repositories/PageRepositoryMock.ts
@@ -1,8 +1,8 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
 import { UpdateWriteOpResult } from 'mongoose';
-import { Page, PageStatus } from '../domains/Page';
-import { PaginationQuery, PaginationOptions } from '../interfaces/pagination';
-import { IPageRepository } from '../repositories/IPageRepository';
+import { Page, PageStatus } from '../../domains/Page';
+import { PaginationQuery, PaginationOptions } from '../../interfaces/pagination';
+import { IPageRepository } from '../../repositories/IPageRepository';
 
 export class PageRepositoryMock implements IPageRepository {
   createPage(page: Partial<Page>): Promise<Page> {
