+++ src/usecases/page/__tests__/MovePageToDirectoryUseCase.test.ts
@@ -1,4 +1,5 @@
-import { generateMockPage, generateMockUser, PageRepositoryMock } from '../../../mock';
+import { generateMockPage, generateMockUser } from '../../../mock/domains';
+import { PageRepositoryMock } from '../../../mock/repositories';
 import { MovePageToDirectoryUseCase } from '../MovePageToDirectoryUseCase';
 
 describe('MovePageToDirectoryUseCase', () => {
